requirements.txt
chromadb>=0.4.0
PyYAML>=6.0
